requirements.txt
# Web framework and server
flask>=2.0.1
gunicorn>=20.1.0
requests>=2.26.0

# Bot Framework
botbuilder-core>=4.14.1
botbuilder-schema>=4.14.1
botframework-connector>=4.14.1
botbuilder-integration-aiohttp>=4.14.1
botbuilder-dialogs>=4.14.1
botbuilder-ai>=4.14.1

# Text-to-speech
pyttsx3>=2.90

# Utilities
python-dotenv>=0.19.2
tqdm>=4.64.1
loguru>=0.6.0 